cast columns to an integer to prevent a deprecation warning from previous commit

+++ ubiquity/components/ubi-language.py
@@ -111,7 +111,7 @@ class PageGtk(PageBase):
             self.iconview.set_model(list_store)
             self.iconview.set_text_column(0)
             lang_per_column = self.iconview.get_allocation().height / 40
-            columns = round(len(choices) / float(lang_per_column))
+            columns = int(round(len(choices) / float(lang_per_column)))
             self.iconview.set_columns(columns)
         else:
             if len(self.treeview.get_columns()) < 1:
